requirements.txt
pandas
numpy
matplotlib
seaborn
yfinance
PyYAML